kernel/scheduler: Implement comprehensive performance and correctness optimizations.
This commit applies a series of enhancements to the Haiku scheduler:
- scheduler_profiler: Use nanosecond precision for internal storage;
  capture CPU ID at entry to ensure stack integrity during migration.
- scheduler_topology: Extend package deduplication bitmask to 512 bits
  and optimize per-CPU index updates using plain stores.
- scheduler_thread: Refactor virtual runtime updates with an optimized
  CAS loop and real-time thread guards; fix race condition in
  _UpdatePriorityBoost by setting fEnqueued=false before removal.
- scheduler_locking: Refactor InterruptGuard to use cpu_status for
  improved correctness and optimize lock-held checks.
- scheduler_common: Add scheduler_atomic_or64 helper and GetCurrentMode
  accessor.
- scheduler_modes: Group configuration constants for better cache
  locality and add update_thread_timeslice hook.
- scheduler_cpu: Simplify idle-node mask transitions using 64-bit
  atomic operations.
- scheduler_tracing: Correct union member syntax, add tracing guards,
  and improve member layout.
- scheduler.cpp: Implement update_thread_timeslice hook call in
  scheduler_on_team_foreground_changed.

+++ src/system/kernel/scheduler/scheduler_tracing.h
@@ -137,8 +137,8 @@ public:
 	virtual const char* Name() const;
 
 	thread_id PreviousThreadID() const		{ return fPreviousID; }
-	uint8 PreviousState() const				{ return (uint8)fPreviousState; }
-	uint16 PreviousWaitObjectType() const	{ return (uint16)fPreviousWaitObjectType; }
+	uint8 PreviousState() const				{ return fPreviousState; }
+	uint16 PreviousWaitObjectType() const	{ return fPreviousWaitObjectType; }
 	const void* PreviousWaitObject() const	{ return fWait.fPreviousWaitObject; }
 
 private:
